OtherPlayerスクリプトを削除し、Playerスクリプトに統一させました。 また、OtherPlayerスクリプトと関連のあるスクリプトを整理しました。
