--- Assets/USERS/Enomoto/02_Scripts/Player/OtherPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0126c239c1b25784d8d4f878ac8a2f3c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
