+++ Assets/USERS/Ueno/02_Scripts/Goldmanager.cs
@@ -37,8 +37,11 @@ public class Goldmanager : MonoBehaviour
     /// <param name="other"></param>
     private void OnTriggerEnter(Collider other)
     {
-        // 加算するスコアをサーバーに送信する関数
-        ScoreMethodList.Instance.SendAddScore();
+        if (EditorManager.Instance.useServer == true)
+        {// サーバーを使用する場合
+            // 加算するスコアをサーバーに送信する関数
+            ScoreMethodList.Instance.SendAddScore();
+        }
 
         Destroy(parentObj);
         Debug.Log("当たった");
